requirements.txt
easyocr==1.6.2
pypinyin==0.47.1
Pillow~=9.3.0